+++ Convert.build.ps1
@@ -77,11 +77,12 @@ Enter-Build {
         $manifestInfo = Test-ModuleManifest -Path $script:ModuleManifestFile
         $script:ModuleVersion = [string]$manifestInfo.Version
         $script:ModuleDescription = $manifestInfo.Description
-        $Script:FunctionsToExport = ($manifestInfo.ExportedCommands.Values | Where-Object {$_.CommandType -eq 'Function'}).Name
+        $Script:FunctionsToExport = ($manifestInfo.ExportedCommands.Values | Where-Object { $_.CommandType -eq 'Function' }).Name
     }
 
     $script:TestsPath = Join-Path -Path $script:SourcePath -ChildPath 'Tests'
     $script:UnitTestsPath = Join-Path -Path $script:TestsPath -ChildPath 'Unit'
+    $script:BuildTestsPath = Join-Path -Path $script:TestsPath -ChildPath 'Build'
     $script:IntegrationTestsPath = Join-Path -Path $script:TestsPath -ChildPath 'Integration'
 
     $script:ArtifactsPath = Join-Path -Path $BuildRoot -ChildPath 'Artifacts'
@@ -220,13 +221,10 @@ task Analyze {
 
     $scriptAnalyzerResults = Invoke-ScriptAnalyzer @scriptAnalyzerParams
 
-    if ($scriptAnalyzerResults)
-    {
+    if ($scriptAnalyzerResults) {
         $scriptAnalyzerResults | Format-Table
         throw 'One or more PSScriptAnalyzer errors/warnings where found.'
-    }
-    else
-    {
+    } else {
         Write-Host '  PowerShell Module Files: Passed' -ForegroundColor Green
     }
     Write-Host ''
@@ -238,8 +236,7 @@ task AnalyzeTests -After Analyze {
     Write-Host '  Pester Test Files: Analyzing...' -ForegroundColor Green
     Write-Host ''
 
-    if (Test-Path -Path $script:TestsPath)
-    {
+    if (Test-Path -Path $script:TestsPath) {
         $scriptAnalyzerParams = @{
             Path        = $script:TestsPath
             ExcludeRule = @(
@@ -254,8 +251,7 @@ task AnalyzeTests -After Analyze {
 
         $scriptAnalyzerResults = Invoke-ScriptAnalyzer @scriptAnalyzerParams
 
-        if ($scriptAnalyzerResults)
-        {
+        if ($scriptAnalyzerResults) {
             $scriptAnalyzerResults | Format-Table
             throw 'One or more PSScriptAnalyzer errors/warnings where found.'
         }
@@ -270,17 +266,17 @@ task Test {
     Write-Host ''
 
     $testPath = $script:UnitTestsPath
-    $codeCoverageFiles = Get-ChildItem -Path $script:ModuleSourcePath -Filter '*.ps1' -Recurse | Where-Object {$_.Name -notlike '_*'} | Select-Object -ExpandProperty FullName
+    $codeCoverageFiles = Get-ChildItem -Path $script:ModuleSourcePath -Filter '*.ps1' -Recurse | Where-Object { $_.Name -notlike '_*' } | Select-Object -ExpandProperty FullName
     if ($TestFile) {
-        $testPath = Get-ChildItem -Path $script:TestsPath -Recurse | Where-Object {$_.Name -like "$TestFile*"} | Select-Object -ExpandProperty FullName
-        $codeCoverageFiles = Get-ChildItem -Path $script:ModuleSourcePath -Filter '*.ps1' -Recurse | Where-Object {$_.Name -like "$TestFile"} | Select-Object -ExpandProperty FullName
+        $testPath = Get-ChildItem -Path $script:TestsPath -Recurse | Where-Object { $_.Name -like "$TestFile*" } | Select-Object -ExpandProperty FullName
+        $codeCoverageFiles = Get-ChildItem -Path $script:ModuleSourcePath -Filter '*.ps1' -Recurse | Where-Object { $_.Name -like "$TestFile" } | Select-Object -ExpandProperty FullName
     }
-    if (-not(Test-Path -Path $testPath)) {return}
+    if (-not(Test-Path -Path $testPath)) { return }
 
     $invokePesterUnitTests = @{
-        Task = 'Test'
-        UnitTestPath = $testPath
-        CodeCoverageFiles = $codeCoverageFiles
+        Task               = 'Test'
+        UnitTestPath       = $testPath
+        CodeCoverageFiles  = $codeCoverageFiles
         EnableCodeCoverage = $true
     }
     InvokePesterUnitTests @invokePesterUnitTests
@@ -347,8 +343,7 @@ task CreateMarkdownHelp {
     $newModuleDocsContent.ToString().TrimEnd() | Out-File -FilePath $ModuleDocsPath -Force -Encoding:utf8
 
     $MissingDocumentation = Select-String -Path (Join-Path -Path $docsPath -ChildPath '\*.md') -Pattern '({{.*}})'
-    if ($MissingDocumentation.Count -gt 0)
-    {
+    if ($MissingDocumentation.Count -gt 0) {
         Write-Host -ForegroundColor Yellow ''
         Write-Host -ForegroundColor Yellow '   The documentation that got generated resulted in missing sections which should be filled out.'
         Write-Host -ForegroundColor Yellow '   Please review the following sections in your comment based help, fill out missing information and rerun this build:'
@@ -437,8 +432,7 @@ task Build {
 
     # TO DO: Add support for Requires Statements by finding them and placing them at the top of the newly created .psm1
     $powerShellScripts = Get-ChildItem -Path $script:ModuleSourcePath -Filter '*.ps1' -Recurse
-    foreach ($script in $powerShellScripts)
-    {
+    foreach ($script in $powerShellScripts) {
         $null = $scriptContent.Append((Get-Content -Path $script.FullName -Raw))
         $null = $scriptContent.AppendLine('')
         $null = $scriptContent.AppendLine('')
@@ -455,6 +449,20 @@ task Build {
     Write-Host ''
 }
 
+task TestBuild -After Build {
+    Write-Host ''
+
+    if (-not(Test-Path -Path $script:BuildTestsPath)) { return }
+
+    $invokePesterUnitTests = @{
+        Task         = 'Build'
+        UnitTestPath = $script:BuildTestsPath
+    }
+    InvokePesterUnitTests @invokePesterUnitTests
+
+    Write-Host ''
+}
+
 # Synopsis: Creates a Module Artifact
 task CreateArtifact {
     Write-Host ''
@@ -462,19 +470,15 @@ task CreateArtifact {
     Write-Host ''
 
     $archivePath = Join-Path -Path $BuildRoot -ChildPath 'Archive'
-    if (Test-Path -Path $archivePath)
-    {
+    if (Test-Path -Path $archivePath) {
         $null = Remove-Item -Path $archivePath -Recurse -Force
     }
 
     $null = New-Item -Path $archivePath -ItemType Directory -Force
 
-    if ($env:CODEBUILD_BUILD_ARN -like '*linux*')
-    {
+    if ($env:CODEBUILD_BUILD_ARN -like '*linux*') {
         $platform = 'linux'
-    }
-    else
-    {
+    } else {
         $platform = 'windows'
     }
 
@@ -490,8 +494,7 @@ task CreateArtifact {
     $script:DeploymentArtifactFileName = '{0}_{1}.zip' -f $script:ModuleName, $script:ModuleVersion
     $script:DeploymentArtifact = Join-Path -Path $script:DeploymentArtifactsPath -ChildPath $script:DeploymentArtifactFileName
 
-    if ($PSEdition -eq 'Desktop')
-    {
+    if ($PSEdition -eq 'Desktop') {
         Add-Type -AssemblyName 'System.IO.Compression.FileSystem'
     }
     [System.IO.Compression.ZipFile]::CreateFromDirectory($script:ArtifactsPath, $script:ZipFile)
@@ -500,15 +503,11 @@ task CreateArtifact {
     Copy-Item -Path $script:ZipFile -Destination $script:DeploymentArtifact
     Write-Host '    Deployment Artifact:  Created' -ForegroundColor Green
 
-    if ($env:CODEBUILD_WEBHOOK_HEAD_REF -and $env:CODEBUILD_WEBHOOK_TRIGGER)
-    {
+    if ($env:CODEBUILD_WEBHOOK_HEAD_REF -and $env:CODEBUILD_WEBHOOK_TRIGGER) {
         Write-Host ('    This was a WebHook triggered build: {0}' -f $env:CODEBUILD_WEBHOOK_TRIGGER)
-        if ($env:CODEBUILD_WEBHOOK_HEAD_REF -eq 'refs/heads/master' -and $env:CODEBUILD_WEBHOOK_TRIGGER -eq 'branch/master')
-        {
+        if ($env:CODEBUILD_WEBHOOK_HEAD_REF -eq 'refs/heads/master' -and $env:CODEBUILD_WEBHOOK_TRIGGER -eq 'branch/master') {
             $s3Bucket = $env:ARTIFACT_BUCKET
-        }
-        else
-        {
+        } else {
             $s3Bucket = $env:DEVELOPMENT_ARTIFACT_BUCKET
         }
 
@@ -517,8 +516,8 @@ task CreateArtifact {
         $s3Key = '{0}/{1}/{2}' -f $script:ModuleName, $branch, $script:ZipFileNameWithPlatform
         $writeS3Object = @{
             BucketName = $s3Bucket
-            Key = $s3Key
-            File = $script:ZipFile
+            Key        = $s3Key
+            File       = $script:ZipFile
         }
         Write-S3Object @writeS3Object
         Write-Host ('    Published artifact to s3://{0}/{1}' -f $s3Bucket, $s3Key)