+++ src/Convert/Convert.psm1
@@ -1,8 +1,7 @@
 $scriptPath = Split-Path $MyInvocation.MyCommand.Path
 try
 {
-    $subFolders = Get-ChildItem -Path $scriptPath -Attributes D
-    $files = Get-ChildItem -Path $subFolders.FullName | Where-Object {$_.Extension -eq '.ps1'}
+    $files = Get-ChildItem -Path $scriptPath -Recurse | Where-Object {$_.Extension -eq '.ps1'}
     foreach ($file in $files)
     {
         . $file.FullName