+++ src/Convert/_RootModule.ps1
@@ -1,0 +1,1 @@
+$script:EPOCH_TIME = Get-Date -Year 1970 -Month 1 -Day 1